- Remove some public APIs from internal headers. - Remove some internal APIs/types from public headers. - Make PSECTION_OBJECT a PVOID unless really needed (which should only be inside \mm) - Make hal buildable on msvc. - Fix some other misc. various msvc/ddk incompatibilities.
svn path=/trunk/; revision=20696

+++ reactos/hal/halx86/generic/beep.c
@@ -50,7 +50,7 @@ HalMakeBeep (
 
     if (Frequency)
     {
-        DWORD Divider = CLOCKFREQ / Frequency;
+        ULONG Divider = CLOCKFREQ / Frequency;
 
         if (Divider > 0x10000)
         {
